refactor: minor fixes for review
